stop complaining

+++ src/managers/InputManager.cpp
@@ -74,7 +74,7 @@ void CInputManager::onMouseButton(wlr_event_pointer_button* e) {
 
     switch (e->state) {
         case WLR_BUTTON_PRESSED:
-            if ((e->button == BTN_LEFT || e->button == BTN_RIGHT) && wlr_keyboard_get_modifiers(PKEYBOARD) == g_pConfigManager->getInt("general:main_mod_internal")) {
+            if ((e->button == BTN_LEFT || e->button == BTN_RIGHT) && wlr_keyboard_get_modifiers(PKEYBOARD) == (uint32_t)g_pConfigManager->getInt("general:main_mod_internal")) {
                 currentlyDraggedWindow = g_pCompositor->windowFloatingFromCursor();
                 dragButton = e->button;
